refactor: consolidate invoice line item tests into a single file
